### 2023-04-20 #5382 @benlansdell
@adi3 Great, thanks for considering this software! Yes, after searching through the list, I found the following reviewers have expertise in pose-tracking/behavior analysis/computational neuroscience. Very roughly in the order of how much their expertise matches:
FelixHenninger
guxiaob
imcatta
Justin Kiggins
Mih
Vjbytes102
