+++ paper.bib
@@ -34,6 +34,7 @@ publisher = {eLife Sciences Publications, Ltd},
   volume={5},
   number={1},
   pages={1267},
+  doi={10.1101/2020.05.14.095430},
   year={2022},
   publisher={Nature Publishing Group UK London}
 }
@@ -81,6 +82,7 @@ doi = {10.1038/s41386-020-0776-y}
     author={Mackenzie W. Mathis and Alexander Mathis},
     journal={Current Opinion in Neurobiology},
     year={2020},
+    doi={10.1016/j.conb.2019.10.008},
     volume={60},
     pages={1-11}
 }
@@ -117,6 +119,7 @@ publisher = {eLife Sciences Publications, Ltd},
   journal={Nature communications},
   volume={12},
   number={1},
+  doi={10.1038/s41467-021-25420-x},
   pages={5188},
   year={2021},
   publisher={Nature Publishing Group UK London}
