### 2023-07-06 #5623 @editorialbot
[REVIEW]: Ethome: machine learning for animal behavior
**Submitting author:** @benlansdell (<a href="http://orcid.org/0000-0003-1444-1950">Benjamin Lansdell</a>)
**Repository:** https://github.com/benlansdell/ethome
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.3
