### 2023-07-07 #5623 @Kevin-Mattheus-Moerman
@neuromusic to get started, the command you were looking for was: `@editorialbot generate my checklist`
