### 2023-07-13 #5623 @imcatta
Hi @benlansdell! I've completed my review. Here are my comments.

The package is interesting and useful. While running through the code, I had the feeling that it is well structured and of good quality. I also appreciated that the package is made available through PiPY and that there is a pipeline for automated testing. However, I think there is still room for improvement in the documentation. Here are my specific comments.

### **General checks**

- The project includes some sample data from animals, but no evidence is provided that this data was collected in accordance with the [JOSS Policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies).

### Functionality

- There are two methods to get example data: `get_sample_nwb__paths()`, which returns NWB data, and `get_sample_data_paths()`, which returns DLC and BORIS. I suggest renaming the latter to something that recalls DLC and BORIS for consistency.

### Documentation

- I struggled a little in following the code presented in the README.
    - `dataset.features.add('cnn1d_prob')` If I run this on the nwb example dataset, as suggested above, it tells me that `DataFrame doesn't have necessary columns to compute this set of features.`. Also, I can’t find which columns are needed. Please consider stating in the documentation what columns are needed and how to get those in the sample dataset.
    - `cross_val_score(..., dataset.ml.labels, ...)`. No labels are present in the nwb example dataset, so this is not working with the dataset proposed above.
    - Please add package requirements (e.g. Python versions supported) to the README.
    - “For this, you must have labeled your body parts in a certain way (refer to How To)”. Please add the link to the How to section.
    - Please add a link to the web documentation in the README.
- It is stated that “`dataset.features.add('cnn1d_prob')`Uses a pretrained CNN to output probabilities of 3 behaviors (attack, mount, social investigation).”. More information on this is needed. How was this CNN built? Also, I found out that the repository contains 5 pre-trained models, but no information on that is provided.
- Some users may not be fully confident with the DLC, SLEAP, NWB, BORIS and NWB file formats. Please explain the different data input formats supported in more detail.
- Guidelines for report issues or problems with the software and seek support are missing.
- In How To guide, section “3a In-built support for resident-intruder setup”, more information on what is MARS and context should be provided.
