### 2023-07-21 benlansdell/ethome#1 @benlansdell
Hi @imcatta, thanks for bringing this to my attention. I will look into what's needed for poetry compatibility. I appreciate the feedback.
