### 2023-10-24 #5623 @benlansdell
@adi3 and @Kevin-Mattheus-Moerman  I am still motivated to take the publication forward at JOSS! Let me address these issues this week. Thanks for keeping this moving.
