### 2023-11-05 75a3f59
Rename get_sample_data function names to mention DLC and BORIS
