+++ docs/index.md
@@ -30,12 +30,12 @@ It's easiest to start with an NWB file, which has metadata already connected to 
 Import
 ```python
 from ethome import create_dataset, create_metadata
-from ethome.io import get_sample_data_paths
+from ethome.io import get_sample_data_paths_dlcboris
 ```
 
 Gather a sample NWB file
 ```python
-tracking_files, annotation_files = get_sample_data_paths()
+tracking_files, annotation_files = get_sample_data_paths_dlcboris()
 ```
 
 Create the dataframe:
