+++ docs/api-docs/README.md
@@ -31,7 +31,7 @@
 - [`interpolation.interpolate_lowconf_points`](./interpolation.md#function-interpolate_lowconf_points): Interpolate raw tracking points if their probabilities are available.
 - [`io.create_behavior_labels`](./io.md#function-create_behavior_labels): Create behavior labels from BORIS exported csv files.
 - [`io.get_sample_data`](./io.md#function-get_sample_data): Load a sample dataset of 5 mice social interaction videos. Each video is approx. 5 minutes in duration
-- [`io.get_sample_data_paths`](./io.md#function-get_sample_data_paths): Get path to sample data files provided with package. 
+- [`io.get_sample_data_paths_dlcboris`](./io.md#function-get_sample_data_paths_dlcboris): Get path to sample data files provided with package. 
 - [`io.get_sample_nwb_paths`](./io.md#function-get_sample_nwb_paths): Get path to a sample NWB file with tracking data for testing and dev purposes.
 - [`io.load_data`](./io.md#function-load_data): Load an object from a pickle file
 - [`io.load_sklearn_model`](./io.md#function-load_sklearn_model): Load sklearn model from file
