### 2023-11-07 19bf159
Add extra info about file formats supported into readme
