### 2023-11-07 #5623 @benlansdell
Hi @adi3, @Kevin-Mattheus-Moerman  and @imcatta, I have made changes based on @imcatta's feedback. The responses to these comments are detailed in Issues #2, #3 in the source repo. I welcome further discussion there about whether they fully satisfy @imcatta's feedback.

One outstanding issue is the use of animal data in the repo. Some of the sample data is from a manuscript in preparation from my previous lab. I assume that because it is as-yet unpublished I need to either:
* Provide all the relevant experimental methods in this repo, OR
* Remove that sample data and replace it with something already published
