### 2023-11-08 #5623 @imcatta
Hi @benlansdell! Thanks for the update. I think the manuscript is moving in the right direction, but there are still a couple of things that can be clarified.

- **Human and animal research:** as you mentioned, this is still an open point at this time
- **Community guidelines:** I'm having trouble locating the guidelines.
