### 2023-11-09 #5623 @benlansdell
@imcatta , The community guidelines are CODE_OF_CONDUCT.md and CONTRIBUTING.md. Let me know if you think something additional is needed. Or if they could be more clearly pointed to in the repo.

For the animal data, I think the best thing is to rework the sample data to remove any unpublished data. Thanks for bringing this point to my attention.
