### 2023-11-14 benlansdell/ethome#5 @benlansdell
Best answer is just to swap out unpublished data for published data/synthetic data and add note to README
