### 2023-11-23 7995bb1
Add statement about use of animal data
