### 2023-11-23 benlansdell/ethome#5 @benlansdell
Added statement about use of data, in [7995bb1](https://github.com/benlansdell/ethome/commit/7995bb13a889f86b77eb61ef331ec1aee35d46e4)
