### 2023-11-23 #5623 @benlansdell
Hi @imcatta, I have added a statement to the readme about the source of the animal data. Let me know if you think this is sufficient details. In particular, the data complies with St Jude's Institutional Animal Care & Use Committee. See here: https://github.com/benlansdell/ethome/issues/5
