### 2023-12-14 #5623 @CBroz1
I'm in the same boat - I wouldn't be able to give this my full attention until the first week of Jan. Lmk if that timeline works for you
