### 2024-01-23 #5623 @KonradDanielewski
Hi @benlansdell and @adi3, I have completed my review.

I have some comments regarding the codebase styling, a minor oversight in the example and a general comment on installation, plus some usage issues I encountered when testing it out:

1. In the example `sample_workflow` it still calls `get_sample_data_paths` instead of `get_sample_data_paths_dlcboris`
2. That may just be a personal preference but I like when a `yaml` file is provided to create a separate environment with `conda` and it contains all the necesarry dependencies - for instance the missing `FFmpeg`
3. In `io.py` `read_DLC_tracks` uses `splitext` but compares to file extensions without the `.` hence if statement is never `True` and cannot load `h5` files.
4. That one is strange, but when running `create_dataset` the `animal_renamer` has to be provided as `None` (even though it's the default value) if not being used, otherwise, at least for me it throws an error like a dict was passed instead of the file path. Happens in `read_boris_annotation` [here](https://github.com/benlansdell/ethome/blob/7995bb13a889f86b77eb61ef331ec1aee35d46e4/ethome/video.py#L765) when loading the `BORIS` csv file.
5. For the `speeds` feature for all my files first five indices are not calculated - which stems from [here](https://github.com/benlansdell/ethome/blob/7995bb13a889f86b77eb61ef331ec1aee35d46e4/ethome/features/generic_features.py#L48) where `dt = features_df['time'].diff(periods = n_shifts)` . Same will happen for `compute_part_speed` [here](https://github.com/benlansdell/ethome/blob/7995bb13a889f86b77eb61ef331ec1aee35d46e4/ethome/features/generic_features.py#L120). Because `periods=5` first five indices have nothing to be compared to resulting in NaNs, propagating them further in the feature extraction. Then modelling crashes as there are NaNs in the data.
6. Lastly, about the styling, this is rather minor and quickly fixable - in some places type hints are provided in other they are not and in some function definitions arguments are in separate lines while in others they are in one line (as far as I noted these are in the features part of the module).
7. More examples could be provided for different inputs so that the users can quickly navigate or follow an example notebook by just pointing to their directories with data.
8. Maybe I missed something in the How to guide but I don't see a clear indication of which features can be added for non-MARS data

Other than that I really like it and think it will be a great resource for many people working with pose estimation data. All the issues are rather minor and fixing them should be quick and easy as the code is very nicely structured and readable.

Let me know if there is anything else you require
