### 2024-01-23 benlansdell/ethome#9 @benlansdell
Fix loading of h5 files
In io.py read_DLC_tracks uses splitext but compares to file extensions without the . hence if statement is never True and cannot load h5 files.
