### 2024-01-23 benlansdell/ethome#10 @benlansdell
create_dataset with animal_renamer has to be provided as none
When running create_dataset the animal_renamer has to be provided as None (even though it's the default value) if not being used, otherwise, at least for me it throws an error like a dict was passed instead of the file path. Happens in read_boris_annotation here when loading the BORIS csv file.
