### 2024-01-23 benlansdell/ethome#12 @benlansdell
type hints
About the styling, this is rather minor and quickly fixable - in some places type hints are provided in other they are not and in some function definitions arguments are in separate lines while in others they are in one line (as far as I noted these are in the features part of the module).
