### 2024-01-23 benlansdell/ethome#13 @benlansdell
More examples
More examples could be provided for different inputs so that the users can quickly navigate or follow an example notebook by just pointing to their directories with data.
