### 2024-01-23 benlansdell/ethome#14 @benlansdell
Better indicate what are non-MARS features
In the How to guide add a clear indication of which features can be added for non-MARS data
