### 2024-01-23 #5623 @benlansdell
Hi @KonradDanielewski, thanks for the review! These are some good points. I have created issues in the ethome repo to track addressing each of these. I'll get started on this.
