### 2024-01-29 964449e
Create conda yaml file, and update readme accordingly
