### 2024-01-29 a4157ca
Update readme to suggest installing tensorflow too. Closes #8
