+++ ethome/features/dl_features.py
@@ -1,5 +1,5 @@
 import numpy as np
-import json 
+import json
 import pandas as pd
 import tqdm
 import os
@@ -12,145 +12,158 @@ from ..utils import check_keras
 
 THIS_FILE_DIR = os.path.dirname(os.path.realpath(__file__))
 
-default_config = {"val_size": 0.2,
-                    "normalize": True,
-                    "split_videos": False,
-                    "future_frames": 50,
-                    "past_frames": 50,
-                    "frame_gap": 1,
-                    "use_conv": True,
-                    "batch_size": 128,
-                    "augment": False,
-                    "epochs": 15,
-                    "reweight_loss": False,
-                    "learning_decay_freq": 1000,
-                    "model_param__layer_channels": [128, 64, 64],
-                    "model_param__conv_size": 5,
-                    "model_param__dropout_rate": 0.5,
-                    "model_param__learning_rate": 0.0005,
-                    "features": "identity"
-                }
+default_config = {
+    "val_size": 0.2,
+    "normalize": True,
+    "split_videos": False,
+    "future_frames": 50,
+    "past_frames": 50,
+    "frame_gap": 1,
+    "use_conv": True,
+    "batch_size": 128,
+    "augment": False,
+    "epochs": 15,
+    "reweight_loss": False,
+    "learning_decay_freq": 1000,
+    "model_param__layer_channels": [128, 64, 64],
+    "model_param__conv_size": 5,
+    "model_param__dropout_rate": 0.5,
+    "model_param__learning_rate": 0.0005,
+    "features": "identity",
+}
 
-feature_spaces = {'distances': [features_distances, [91]], 
-                    'identity': [features_identity, [2,7,2]],
-                    'distances_normalized': [features_distances_normalized, [91]],
-                    'mars': [features_mars, []],
-                    'mars_distr': [features_mars_distr, []],
-                }
-                    # This one was removed....I think it's not needed
-                    # 'mars_no_shift': [features_mars_no_shift, [160]]
+feature_spaces = {
+    "distances": [features_distances, [91]],
+    "identity": [features_identity, [2, 7, 2]],
+    "distances_normalized": [features_distances_normalized, [91]],
+    "mars": [features_mars, []],
+    "mars_distr": [features_mars_distr, []],
+}
+# This one was removed....I think it's not needed
+# 'mars_no_shift': [features_mars_no_shift, [160]]
 
 sweeps_baseline = {
-    'test_run_distances':
-        [{'num_samples': 1},
+    "test_run_distances": [
+        {"num_samples": 1},
         {
             "model_param__learning_rate": 0.0001,
             "epochs": 15,
             "features": "distances",
-            "model_param__dropout_rate": 0.5
-        }]
+            "model_param__dropout_rate": 0.5,
+        },
+    ]
 }
 
-def seed_everything(seed = 2012):
+
+def seed_everything(seed=2012):
     np.random.seed(seed)
-    os.environ['PYTHONHASHSEED'] = str(seed)
+    os.environ["PYTHONHASHSEED"] = str(seed)
+
 
 seed_everything()
 
-#Implement cross validation, instead of a single validation set
-#Product predictions on the held out data, all merged into one frame
-class Trainer(object):
-    def __init__(self, *,
-               feature_dim, 
-               num_classes,
-               test_data = None,
-               class_to_number=None,
-               past_frames=0, 
-               future_frames=0,
-               frame_gap=1, 
-               use_conv=False, 
-               build_model = build_baseline_model,
-               Generator = MABe_Generator,
-               use_callbacks = False,
-               learning_decay_freq = 10,
-               featurizer = features_identity):
 
+# Implement cross validation, instead of a single validation set
+# Product predictions on the held out data, all merged into one frame
+class Trainer(object):
+    def __init__(
+        self,
[…]
