+++ ethome/features/features.py
@@ -4,33 +4,62 @@
 import warnings
 
 from ethome.features.dl_features import compute_dl_probability_features
-from ethome.features.mars_features import compute_mars_features, compute_mars_reduced_features, compute_social_features
-from ethome.features.generic_features import compute_centerofmass_interanimal_distances, \
-                                        compute_centerofmass_interanimal_speed, \
-                                        compute_centerofmass, \
-                                        compute_centerofmass_velocity, \
-                                        compute_speed_features, \
-                                        compute_distance_features
+from ethome.features.mars_features import (
+    compute_mars_features,
+    compute_mars_reduced_features,
+    compute_social_features,
+)
+from ethome.features.generic_features import (
+    compute_centerofmass_interanimal_distances,
+    compute_centerofmass_interanimal_speed,
+    compute_centerofmass,
+    compute_centerofmass_velocity,
+    compute_speed_features,
+    compute_distance_features,
+)
 
-default_tracking_columns = ['resident_x_nose', 'resident_x_leftear', 'resident_x_rightear', 'resident_x_neck',
-                            'resident_x_lefthip', 'resident_x_righthip', 'resident_x_tail', 'resident_y_nose',
-                            'resident_y_leftear', 'resident_y_rightear', 'resident_y_neck',
-                            'resident_y_lefthip', 'resident_y_righthip', 'resident_y_tail',
-                            'intruder_x_nose', 'intruder_x_leftear', 'intruder_x_rightear',
-                            'intruder_x_neck', 'intruder_x_lefthip', 'intruder_x_righthip',
-                            'intruder_x_tail', 'intruder_y_nose', 'intruder_y_leftear',
-                            'intruder_y_rightear', 'intruder_y_neck', 'intruder_y_lefthip',
-                            'intruder_y_righthip', 'intruder_y_tail']
+default_tracking_columns = [
+    "resident_x_nose",
+    "resident_x_leftear",
+    "resident_x_rightear",
+    "resident_x_neck",
+    "resident_x_lefthip",
+    "resident_x_righthip",
+    "resident_x_tail",
+    "resident_y_nose",
+    "resident_y_leftear",
+    "resident_y_rightear",
+    "resident_y_neck",
+    "resident_y_lefthip",
+    "resident_y_righthip",
+    "resident_y_tail",
+    "intruder_x_nose",
+    "intruder_x_leftear",
+    "intruder_x_rightear",
+    "intruder_x_neck",
+    "intruder_x_lefthip",
+    "intruder_x_righthip",
+    "intruder_x_tail",
+    "intruder_y_nose",
+    "intruder_y_leftear",
+    "intruder_y_rightear",
+    "intruder_y_neck",
+    "intruder_y_lefthip",
+    "intruder_y_righthip",
+    "intruder_y_tail",
+]
 
-class Features: # pragma: no cover
+
+class Features:  # pragma: no cover
     def __init__(self):
         raise NotImplementedError
 
     def transform(self, df):
         raise NotImplementedError
 
-def feature_class_maker(name, compute_function, required_columns = []):
-    def __init__(self, required_columns = None, **kwargs):
+
+def feature_class_maker(name, compute_function, required_columns=[]):
+    def __init__(self, required_columns=None, **kwargs):
         """Feature creation object. This houses the feature creation function and the columns that are required to compute the features. Performs some checks on data to make sure has these columns.
 
         See docstring for the `features` model for more information.
@@ -42,8 +71,8 @@ def feature_class_maker(name, compute_function, required_columns = []):
             self.required_columns = required_columns
         self.kwargs = kwargs
 
[…]
