+++ tests/test_models.py
@@ -2,11 +2,8 @@
 ## Setup test code ##
 #####################
 
-import pytest
-from ethome import create_metadata, interpolate_lowconf_points, video
-import pandas as pd
 
-#Metadata is a dictionary
+# Metadata is a dictionary
 # def test_f1_optimizer(videodataset_mars):
 #     """ Test creation of metadata object """
 
