+++ paper.bib
@@ -18,7 +18,7 @@ issn = {2050-084X},
 publisher = {eLife Sciences Publications, Ltd},
 }
 @misc{Sun2020-TREBA,
-  doi = {10.48550/ARXIV.2011.13917},  
+  doi = {10.48550/ARXIV.2011.13917},
   url = {https://arxiv.org/abs/2011.13917},
   author = {Sun, Jennifer J. and Kennedy, Ann and Zhan, Eric and Anderson, David J. and Yue, Yisong and Perona, Pietro},
   keywords = {Computer Vision and Pattern Recognition (cs.CV), Machine Learning (cs.LG), FOS: Computer and information sciences, FOS: Computer and information sciences},
@@ -58,7 +58,7 @@ publisher = {eLife Sciences Publications, Ltd},
                   Hausmann, Sébastien and
                   Schneider, Steffen and
                   Mathis, Alexander},
-  title        = {{DLC2Kinematics: a post-deeplabcut module for 
+  title        = {{DLC2Kinematics: a post-deeplabcut module for
                    kinematic analysis}},
   month        = feb,
   year         = 2020,
