+++ paper.md
@@ -15,7 +15,7 @@ authors:
     equal-contrib: false
     affiliation: 1
   - name: Abbas Shirinifard
-    equal-contrib: false 
+    equal-contrib: false
     affiliation: 1
 affiliations:
  - name: Developmental Neurobiology, St Jude Children's Research Hospital, Memphis, Tennessee, USA
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-`ethome` supports machine learning of animal behavior. It interprets pose-tracking files and behavior annotations to create feature tables, train behavior classifiers, interpolate pose tracking data and other common analysis tasks. 
+`ethome` supports machine learning of animal behavior. It interprets pose-tracking files and behavior annotations to create feature tables, train behavior classifiers, interpolate pose tracking data and other common analysis tasks.
 It features:
 
 * Read in animal pose data and corresponding behavior annotations to make supervised learning easy
@@ -37,30 +37,30 @@ It features:
 * Create features specifically for mouse resident-intruder setup
 * Quickly generate a movie with behavior predictions
 
-Together these features significantly reduce the code a user needs to write to perform animal behavior analysis -- they can focus exclusively on developing the best machine learning models for their problem, instead of dealing with data imports, unit conversations, time alignments between data sources, etc. The current version supports DeepLabCut (DLC) and SLEAP [@Pereira2022sleap] tracking data, BORIS behavioral annotations, and NWB behavioral data files, all heavily used standards. The demo notebooks provided in this repository demonstrate how few lines of code are needed to create useful ML models. 
+Together these features significantly reduce the code a user needs to write to perform animal behavior analysis -- they can focus exclusively on developing the best machine learning models for their problem, instead of dealing with data imports, unit conversations, time alignments between data sources, etc. The current version supports DeepLabCut (DLC) and SLEAP [@Pereira2022sleap] tracking data, BORIS behavioral annotations, and NWB behavioral data files, all heavily used standards. The demo notebooks provided in this repository demonstrate how few lines of code are needed to create useful ML models.
 
 # Statement of need
 
-The quantitative study animal behavior is rapidly growing [@Mathis2020DeepLT], with pose-tracking tools like DeepLabCut [@Lauer2022MultianimalPE] making accurate pose estimation easily and widely applicable. Such data promises to revolutionize the study of animal behavior. 
+The quantitative study animal behavior is rapidly growing [@Mathis2020DeepLT], with pose-tracking tools like DeepLabCut [@Lauer2022MultianimalPE] making accurate pose estimation easily and widely applicable. Such data promises to revolutionize the study of animal behavior.
 
-With the vast amount of data available, there is a growing need for analysis tools that simplify post-tracking machine learning tasks. After poses have been tracked in a series of videos, both supervised learning (behavior classification) and unsupervised learning can offer valuable insights. Currently, there are two main options for conducting machine learning analysis: using a comprehensive behavior analysis software package like SimBA [@Nilsson2020-simba], or writing custom analysis code. 
- 
-SimBA provides a user-friendly GUI, making it an ideal solution for those who prefer a no-code approach. However, this convenience comes at the cost of flexibility. On the other hand, writing custom analysis code requires navigating numerous tedious details, such as understanding the file formats used by the software generating the behavioral data, aligning pose and behavior annotation data, converting units, interpolating or correcting inaccurate pose estimates, and creating relevant features for machine learning. This process can be quite time-consuming. 
[…]
