+++ docs/api-docs/plot.md
@@ -27,22 +27,22 @@ plot_embedding(
 ) → tuple
 ```
 
-Scatterplot of a 2D TSNE or UMAP embedding from the dataset. 
+Scatterplot of a 2D TSNE or UMAP embedding from the dataset.
 
 
 
 **Args:**
- 
- - <b>`dataset`</b>:  data 
- - <b>`col_names`</b>:  list of column names to use for the x and y axes 
- - <b>`color_col`</b>:  if provided, a column that will be used to color the points in the scatter plot 
- - <b>`figsize`</b>:  tuple with the dimensions of the plot (in inches) 
- - <b>`kwargs`</b>:  All other keyword pairs are sent to Matplotlib's scatter function 
+
+ - <b>`dataset`</b>:  data
+ - <b>`col_names`</b>:  list of column names to use for the x and y axes
+ - <b>`color_col`</b>:  if provided, a column that will be used to color the points in the scatter plot
+ - <b>`figsize`</b>:  tuple with the dimensions of the plot (in inches)
+ - <b>`kwargs`</b>:  All other keyword pairs are sent to Matplotlib's scatter function
 
 
 
 **Returns:**
- tuple (fig, axes). The Figure and Axes objects.  
+ tuple (fig, axes). The Figure and Axes objects.
 
 
 ---
@@ -61,22 +61,22 @@ plot_unsupervised_results(
 )
 ```
 
-Set of plots for unsupervised behavior clustering results 
+Set of plots for unsupervised behavior clustering results
 
 
 
 **Args:**
- 
- - <b>`dataset`</b>:  data 
- - <b>`cluster_results`</b>:  tuple output by 'cluster_behaviors' 
- - <b>`col_names`</b>:  list of column names to use for the x and y axes 
- - <b>`figsize`</b>:  tuple with the plot dimensions, in inches 
- - <b>`kwargs`</b>:  all other keyword pairs are sent to Matplotlib's scatter function 
+
+ - <b>`dataset`</b>:  data
+ - <b>`cluster_results`</b>:  tuple output by 'cluster_behaviors'
+ - <b>`col_names`</b>:  list of column names to use for the x and y axes
+ - <b>`figsize`</b>:  tuple with the plot dimensions, in inches
+ - <b>`kwargs`</b>:  all other keyword pairs are sent to Matplotlib's scatter function
 
 
 
 **Returns:**
- tuple (fig, axes). The Figure and Axes objects.  
+ tuple (fig, axes). The Figure and Axes objects.
 
 
 ---
@@ -95,21 +95,21 @@ plot_ethogram(
 ) → tuple
 ```
 
-Simple ethogram of one video, up to a certain frame number. 
+Simple ethogram of one video, up to a certain frame number.
 
 
 
 **Args:**
-  dataset:  
- - <b>`vid_key`</b>:  key (in dataset.metadata) pointing to the video to make ethogram for 
- - <b>`query_label`</b>:  the column containing the behavior labels to plot 
- - <b>`frame_limit`</b>:  only make the ethogram for frames between [0, frame_limit] 
- - <b>`figsize`</b>:  tuple with figure size (in inches) 
+  dataset:
+ - <b>`vid_key`</b>:  key (in dataset.metadata) pointing to the video to make ethogram for
+ - <b>`query_label`</b>:  the column containing the behavior labels to plot
+ - <b>`frame_limit`</b>:  only make the ethogram for frames between [0, frame_limit]
+ - <b>`figsize`</b>:  tuple with figure size (in inches)
 
 
 
 **Returns:**
- tuple (fig, axes). The Figure and Axes objects 
[…]
