### 2024-01-31 8189f1a
Use get_sample_data_paths_dlcboris in sample_workflow.ipynb. Closes #7
