### 2024-01-31 e7a0dac
Small changes to how to guide to emphasize which features work with non MARS setup
