### 2024-01-31 9d01112
Small changes to how to guide to emphasize which features work with non MARS setup
