### 2024-01-31 2b625c2
Small changes to how to guide to emphasize which features work with non MARS setup. Closes #14
