### 2024-02-01 9327a4c
Fix bug in extracting extension from filename -- was never loading h5 files. Closes #9
