### 2024-02-02 40b76a6
Add bfill to all diff and shift to avoid creation of nans in first few rows. Closes #11
