### 2024-02-02 3812c08
Start to add type hints (as per #12)
