### 2024-02-02 4a03342
More type hints (#12)
