+++ ethome/io.py
@@ -2,6 +2,7 @@
 import pandas as pd
 import pickle
 import numpy as np
+from typing import Sequence
 from itertools import product
 from joblib import dump, load
 import os
@@ -15,14 +16,14 @@ XY_IDS = ["x", "y"]
 XYLIKELIHOOD_IDS = ["x", "y", "likelihood"]
 
 
-def uniquifier(seq):
+def uniquifier(seq:Sequence):
     """Return a sequence (e.g. list) with unique elements only, but maintaining original list order"""
     seen = set()
     seen_add = seen.add
     return [x for x in seq if not (x in seen or seen_add(x))]
 
 
-def _list_replace(ls, renamer):
+def _list_replace(ls:list, renamer:dict):
     """Replace elements in a list according to provided dictionary"""
     for i, word in enumerate(ls):
         if word in renamer.keys():
@@ -30,7 +31,7 @@ def _list_replace(ls, renamer):
     return ls
 
 
-def save_sklearn_model(model, fn_out):  # pragma: no cover
+def save_sklearn_model(model, fn_out:str):  # pragma: no cover
     """Save sklearn model to file
 
     Args:
@@ -40,7 +41,7 @@ def save_sklearn_model(model, fn_out):  # pragma: no cover
     dump(model, fn_out)
 
 
-def load_sklearn_model(fn_in):  # pragma: no cover
+def load_sklearn_model(fn_in: str):  # pragma: no cover
     """Load sklearn model from file
 
     Args:
@@ -152,7 +153,7 @@ def read_sleap_tracks(
 # POSSIBILITY OF SUCH DAMAGE.
 
 
-def _load_sleap_data(path, multi_animal=False):
+def _load_sleap_data(path:str, multi_animal:bool=False):
     """loads sleap data h5 format from file path and returns it as pd.DataFrame
     As sleap body parts (nodes) are not ordered in a particular way, we sort them alphabetically.
     As sleap tracks do not export a score/likelihood but cut off automatically (nan), we are simulating a likelihood
@@ -408,7 +409,7 @@ def _read_DLC_tracks(
 # SOFTWARE.
 
 
-def _convert_nwb_to_h5_all(nwbfile):
+def _convert_nwb_to_h5_all(nwbfile:str):
     """
     Convert a NWB data file back to DeepLabCut's h5 data format.
 
@@ -546,7 +547,7 @@ def load_data(fn: str):  # pragma: no cover
 
 
 # Only used to making a test dataframe for testing and dev purposes
-def _make_sample_dataframe(fn_out="sample_dataframe.pkl"):  # pragma: no cover
+def _make_sample_dataframe(fn_out:str="sample_dataframe.pkl"):  # pragma: no cover
     from ethome import create_dataset, create_metadata
 
     cur_dir = os.path.dirname(os.path.abspath(__file__))
@@ -662,7 +663,7 @@ def read_boris_annotation(
     return ground_truth
 
 
-def create_behavior_labels(boris_files):
+def create_behavior_labels(boris_files:list):
     """Create behavior labels from BORIS exported csv files.
 
     Args:
