### 2024-02-02 d4cd002
Add type hints to video.py (#12)
