+++ ethome/video.py
@@ -8,6 +8,7 @@ import re
 import warnings
 import types
 
+from typing import Any
 from glob import glob
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.model_selection import cross_val_predict, LeaveOneGroupOut, PredefinedSplit
@@ -49,12 +50,12 @@ UNIT_DICT = {
 }
 
 
-def _add_item_to_dict(tracking_files, metadata, k, item):
+def _add_item_to_dict(tracking_files: list, metadata:dict, k:Any, item:Any):
     for fn in tracking_files:
         metadata[fn][k] = item
 
 
-def _add_items_to_dict(tracking_files, metadata, k, items):
+def _add_items_to_dict(tracking_files: list, metadata:dict, k:Any, items:Any):
     for fn, item in zip(tracking_files, items):
         metadata[fn][k] = item
 
@@ -95,7 +96,7 @@ def create_metadata(tracking_files: list, **kwargs) -> dict:
     return metadata
 
 
-def _convert_units(df):
+def _convert_units(df:pd.DataFrame):
     # if 'frame_width', 'resolution' and 'frame_width_units' are provided, then we convert tracks to these units.
     if len(df.metadata.details) == 0:
         return
@@ -110,7 +111,7 @@ def _convert_units(df):
     df.drop(columns="scale_factor", inplace=True)
 
 
-def _validate_metadata(metadata, req_cols):
+def _validate_metadata(metadata:dict, req_cols:list):
     has_all_dim_cols_count = 0
 
     should_rescale = None
@@ -215,7 +216,7 @@ def _validate_metadata(metadata, req_cols):
 
 @pd.api.extensions.register_dataframe_accessor("metadata")
 class EthologyMetadataAccessor(object):
-    def __init__(self, pandas_obj):
+    def __init__(self, pandas_obj:pd.DataFrame):
         self._obj = pandas_obj
         if "metadata__details" not in self._obj.attrs:
             self._obj.attrs["metadata__details"] = {}
@@ -231,11 +232,11 @@ class EthologyMetadataAccessor(object):
         return self._obj.attrs["metadata__label_key"]
 
     @details.setter
-    def details(self, val):
+    def details(self, val:Any):
         self._obj.attrs["metadata__details"] = val
 
     @label_key.setter
-    def label_key(self, val):
+    def label_key(self, val:Any):
         self._obj.attrs["metadata__label_key"] = val
 
     @property
@@ -253,7 +254,7 @@ class EthologyMetadataAccessor(object):
 
 @pd.api.extensions.register_dataframe_accessor("features")
 class EthologyFeaturesAccessor(object):
-    def __init__(self, pandas_obj):
+    def __init__(self, pandas_obj:pd.DataFrame):
         self._obj = pandas_obj
         if "features__active" not in self._obj.attrs:
             self._obj.attrs["features__active"] = None
@@ -263,7 +264,7 @@ class EthologyFeaturesAccessor(object):
         return self._obj.attrs["features__active"]
 
     @active.setter
-    def active(self, val):
+    def active(self, val:Any):
         self._obj.attrs["features__active"] = val
 
     def activate(self, name: str) -> list:
@@ -315,7 +316,7 @@ class EthologyFeaturesAccessor(object):
     # Set features by individual or by group names
     def add(
         self,
-        feature_maker,
+        feature_maker: Any,
         featureset_name: str = None,
         add_to_features=True,
         required_columns=[],
@@ -395,7 +396,7 @@ class EthologyFeaturesAccessor(object):
 
 @pd.api.extensions.register_dataframe_accessor("pose")
[…]
