### 2024-02-02 benlansdell/ethome#12 @benlansdell
Used `ruff format` to make this more consistent. Added type hints to all function arguments.
