### 2024-02-02 #5623 @benlansdell
Thanks again @KonradDanielewski for providing feedback. I've addressed almost all of the issues you raised. One remaining one is listed here: https://github.com/benlansdell/ethome/issues/10

It's the issue you pointed out with having to provide `animal_renamer = None`, even though it's already None by default. I can't seem to reproduce this one. Are you able to provide some short code that produces the error? (Possibly with a BORIS file, if you used one that's not included in the sample data in the package) Thanks!
