### 2024-02-04 #5623 @KonradDanielewski
> It's the issue you pointed out with having to provide `animal_renamer = None`, even though it's already None by default. I can't seem to reproduce this one.

Hi, so I tried to replicate it and also couldn't... Maybe let's consider this addressed. If I encounter it again I will log what happened exactly and make an issue - this could've also been some weird error on my part as a I was applying quickfixes to the code while testing to provide explanations
