### 2024-02-08 04851ba
Add sample notebooks for NWB and SLEAP analyses. Closes #13
