+++ ethome/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.6.0"
+__version__ = "0.6.1"
 
 # Suppress tensorflow import text...
 import os
