+++ ethome/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.6.1"
+__version__ = "0.6.2"
 
 # Suppress tensorflow import text...
 import os
