+++ docs/index.md
@@ -1,6 +1,7 @@
 [![codecov](https://codecov.io/gh/benlansdell/ethome/branch/master/graph/badge.svg?token=IJ0JJBOGGS)](https://codecov.io/gh/benlansdell/ethome)
 ![build](https://github.com/benlansdell/ethome/actions/workflows/workflow.yml/badge.svg)
 [![PyPI version](https://badge.fury.io/py/ethome-ml.svg)](https://badge.fury.io/py/ethome-ml)
+[![status](https://joss.theoj.org/papers/0472dab158806827a83da79e602e16e4/status.svg)](https://joss.theoj.org/papers/0472dab158806827a83da79e602e16e4)
 
 # Ethome
 
@@ -9,6 +10,8 @@ Tools for machine learning of animal behavior.
 This library interprets pose-tracking files and behavior annotations to create features, train behavior classifiers, interpolate pose tracking data and other common analysis tasks.
 
 At present pose tracking data from DLC, SLEAP and NWB formats are supported, and behavior annotations from BORIS and NWB formats are supported.
+
+Full documentation is posted here: [https://benlansdell.github.io/ethome/](https://benlansdell.github.io/ethome/).
 
 ## Features
 
@@ -25,19 +28,36 @@ At present pose tracking data from DLC, SLEAP and NWB formats are supported, and
 pip install ethome-ml
 ```
 
+`ethome` has been tested with Python 3.7 and 3.8.
+
+### Conda environment
+
+Note that dependencies have tried to be kept to a minimum so that `ethome` can work easily alongside other programs that may be part of your behavior analysis pipeline (e.g. `DeepLabCut`) -- thus you can try running the `pip install` line above in an existing virtual environment.
+
+That said, you may want a separate environment for running `ethome`. A conda environment can be created with the following steps:
+
+1. Download the conda environment yaml file [ethome-conda.yaml](www.google.com)
+2. (From the location you downloaded the yaml file) Create the environment: `conda env create -f ethome-conda.yaml`
+3. Run `conda activate ethome`
+3. And finally `pip install ethome-ml`
+
+### Optional packages
+
+With both install methods, you may want to also install `tensorflow` (at least version 2.0) if you want to use the CNN features for a resident-intruder setup.
+
 ## Quickstart
 
 It's easiest to start with an NWB file, which has metadata already connected to the pose data.
 
 Import
 ```python
-from ethome import create_dataset, create_metadata
-from ethome.io import get_sample_data_paths_dlcboris
+from ethome import create_dataset
+from ethome.io import get_sample_nwb_paths
 ```
 
 Gather a sample NWB file
 ```python
-tracking_files, annotation_files = get_sample_data_paths_dlcboris()
+fn_in = get_sample_nwb_paths()
 ```
 
 Create the dataframe:
@@ -55,11 +75,18 @@ dataset.features.add('distances')
 ```
 will compute all distances between all body parts (both between and within animals).
 
-There are featuresets specifically tailored for social mice studies (resident intruder). For instance,
+We can load pose data from DLC, and behavior annotation data from BORIS, provided we also provide a little metadata for context. E.g.:
+```python
+pose_files, behavior_files = get_sample_data_paths_dlcboris()
+metadata = create_metadata(pose_files, labels = behavior_files, fps = 30)
+dataset = create_dataset(metadata)
+```
+
+There are featuresets specifically tailored for social mice studies (the resident-intruder setup). For instance,
 ```python
 dataset.features.add('cnn1d_prob')
 ```
-Uses a pretrained CNN to output probabilities of 3 behaviors (attack, mount, social investigation). For this, you must have labeled your body parts in a certain way (refer to How To). Other, more generic, feature creation functions are provided that work for any animal configuration.
[…]
