### 2024-02-08 #5623 @benlansdell
Hi @KonradDanielewski,

Ok, good to know. I closed the issue. If it comes up again, I'll reopen and investigate. I believe I've addressed the points you raised. If you're curious, the issues were tracked (and closed with corresponding commits), as follows:

* [Better indicate what are MARS features](https://github.com/benlansdell/ethome/issues/14)
* [More sample workflows](https://github.com/benlansdell/ethome/issues/13)
* [Type hints](https://github.com/benlansdell/ethome/issues/12)
* [Nans from computed speed feature propagate into feature creation](https://github.com/benlansdell/ethome/issues/11)
* [Fix loading of h5 files](https://github.com/benlansdell/ethome/issues/9)
* [Create yaml file for making a conda environment](https://github.com/benlansdell/ethome/issues/8)
* [sample_workflow calls get_sample_data_paths instead of get_sample_data_paths_dlcboris](https://github.com/benlansdell/ethome/issues/7)

Let me know if you have further suggestions, or things you've noticed I may have missed in the above changes. Thanks.
