+++ ethome/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.6.2"
+__version__ = "0.6.3"
 
 # Suppress tensorflow import text...
 import os
