### 2024-02-19 #5623 @benlansdell
@adi3 great! Here is the DOI: 10.5281/zenodo.10680136 Archived on zenodo. Version 0.6.3
