### 2024-02-27 #5623 @Kevin-Mattheus-Moerman
@adi3 next time you can call `@editorialbot create post-review checklist` to create the above. This will help us to ensure that the archive title, license etc. matches and all the final checks are done. I'll take it from here though.
