### 2024-02-27 #5623 @Kevin-Mattheus-Moerman
@benlansdell I have checked the paper,  your repository, this review, and the archive link. Most seems in order. However the below points require your attention:

- [ ] Ensure the archive link title matches the paper title.
- [ ] Ensure the archive listed license matches the software license.
- [ ] Ensure the archive listed author set matches the paper author set.
- [ ] In your paper affiliations, please spell out `United States of America`
- [ ] Add the following DOI's:
* [10.1038/s41592-022-01443-0](https://doi.org/10.1038/s41592-022-01443-0)
* [10.1038/s41592-022-01426-1](https://doi.org/10.1038/s41592-022-01426-1)
