+++ paper.bib
@@ -92,7 +92,8 @@ doi = {10.1038/s41386-020-0776-y}
     journal={Nature Methods},
     year={2022},
     volume={19},
-    pages={496 - 504}
+    pages={496 - 504},
+    doi={10.1038/s41592-022-01443-0}
 }
 @article {Segalin2021-bento,
 article_type = {journal},
@@ -163,6 +164,7 @@ publisher = {eLife Sciences Publications, Ltd},
    volume={19},
    number={4},
    year={2022},
-   publisher={Nature Publishing Group}
+   publisher={Nature Publishing Group},
+   doi={10.1038/s41592-022-01426-1}
    }
 }
