### 2024-02-28 #5623 @benlansdell
@Kevin-Mattheus-Moerman I've corrected the title, license and author list in the Zenodo archive so it matches the paper data, along with the DOIs and affiliation. Let me know if there's anything else.
