### 2024-03-08 e3f7d42
Create CITATION.cff
