+++ CITATION.cff
@@ -1,0 +1,30 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Lansdell
+  given-names: Benjamin
+  orcid: "https://orcid.org/0000-0003-1444-1950"
+- family-names: Shirinifard
+  given-names: Abbas
+doi: 10.5281/zenodo.10680136
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Lansdell
+    given-names: Benjamin
+    orcid: "https://orcid.org/0000-0003-1444-1950"
+  - family-names: Shirinifard
+    given-names: Abbas
+  date-published: 2024-03-08
+  doi: 10.21105/joss.05623
+  issn: 2475-9066
+  issue: 95
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5623
+  title: "Ethome: tools for machine learning of animal behavior"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05623"
+  volume: 9
+title: "Ethome: tools for machine learning of animal behavior"
